requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
beautifulsoup4>=4.9.3
html2text>=2020.1.16
jinja2>=3.0.0
python-dotenv>=0.19.0
resend>=0.5.0